simplify language, typos, tweaks
in addition to the read-through for simplifying language and catching typos...

get healthy again also needed to emphasize rest. added it as a first point. also added under keep moving, don't if you're too tired. been hearing / reading more lately about fatigue is a common symptom and people should rest early and often with this... so it seemed missing and important to highlight here.