update anosmia
loss of smell and taste - we had it as later in the infection, but new research https://www.washingtonpost.com/health/coronavirus-destroys-lungs-but-doctors-are-finding-its-damage-in-kidneys-hearts-and-elsewhere/2020/04/14/7ff71ee0-7db1-11ea-a3ee-13e1ae0a3571_story.html 

suggests it can be an early symptom as well. 

this also seems to be true of the skin symptoms, so collapsed those a bit at end of symptoms description/list paragraph.